Move FilmControls component & getFilmControlsData() to folder
